requirements.txt
aiohttp==3.6.2
aiowinreg==0.0.3
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.10
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.5.3
buttons==0.1.9
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
Cybernator==0.6.3
discord==1.0.1
discord.py==1.6.0
dnspython==2.1.0
emoji==1.2.0
emojis==0.6.0
Flask==1.1.2
future==0.18.2
geojson==2.5.0
gitdb==4.0.5
gitdb2==4.0.2
GitPython==2.1.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.2
MarkupSafe==1.1.1
motor==2.3.0
multidict==4.7.6
Nuitka==0.6.11.4
olefile==0.46
pefile==2019.4.18
Pillow==8.1.0
psutil==5.8.0
pyfiglet==0.7.5
PyGithub==1.34
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyJWT==2.0.1
pymongo==3.11.2
PySocks==1.7.1
pytz==2017.2
pywin32-ctypes==0.2.0
requests==2.25.1
smmap==3.0.5
strawpy==1.1
Tokage==1.0.0
typing-extensions==3.7.4.3
urllib3==1.26.3
websockets==5.0.1
Werkzeug==1.0.1
yarl==1.6.3
you==0.5.0.dev1